Change int on size_t

+++ dijkstra/typedef.h
@@ -2,11 +2,19 @@
 #define TYPEDEF_H
 
 #include <vector>
+#include <sys/types.h>
+#include <string>
+#include <list>
+#include <set>
+#include <cmath>
+#include <QXmlStreamReader>
+#include <QFile>
 
-typedef int vertex_t;
+typedef ssize_t vertex_t;
 typedef double weight_t;
 
 struct neighbor {
+
     vertex_t target;
     weight_t weight;
     neighbor(vertex_t arg_target, weight_t arg_weight)
